% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-output.R
\name{all_output}
\alias{all_output}
\title{Title}
\usage{
all_output(p, image_name, save = TRUE)
}
\arguments{
\item{p}{ggplot2 object}

\item{image_name}{string}

\item{save}{logical}
}
\value{
some plots! saved where you want them
}
\description{
Title
}
