% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buy-list-add.R
\name{buy_list_add}
\alias{buy_list_add}
\title{Add securities to Buy List}
\usage{
buy_list_add(addition, type = "equity")
}
\arguments{
\item{addition}{the ticker of the security to add}

\item{type}{the type of security}
}
\value{
none. writes to file
}
\description{
Add securities to Buy List
}
