% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative-returns.R
\name{comparative_return}
\alias{comparative_return}
\title{Title}
\usage{
comparative_return(ticker, sector, return = "total")
}
\arguments{
\item{ticker}{a character. Ticker from buy list}

\item{sector}{a character. GICS sector of ticker security}

\item{return}{a character. Type of return to use (price or total)}
}
\value{
a table
}
\description{
Title
}
