% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(
  ticker,
  path = "data.xlsx",
  flds = "PX_LAST",
  start_date = "2000-01-01",
  names = NA
)
}
\arguments{
\item{ticker}{character}

\item{path}{character}

\item{flds}{character}

\item{start_date}{character}

\item{names}{character}
}
\value{

}
\description{
Retrieves data from either a .xlsx or Bloomberg API and returns a tibble
}
\examples{

}
