% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path.R
\name{get_path}
\alias{get_path}
\title{Get path to PAM Research Dropbox path}
\usage{
get_path()
}
\value{
A character string to the PAM Research Dropbox path specific to machine
}
\description{
Get path to PAM Research Dropbox path
}
