% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-recommendations.R
\name{get_recommendations}
\alias{get_recommendations}
\title{Title}
\usage{
get_recommendations(ticker = NA)
}
\arguments{
\item{ticker}{if NA returns the entire recommendations dataset. If ticker is given returns recommendations for the desired security}
}
\value{
a tibble
}
\description{
Title
}
