% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last-recommendation.R
\name{last_recommendation}
\alias{last_recommendation}
\title{Get the most recent recommendation}
\usage{
last_recommendation(ticker)
}
\arguments{
\item{ticker}{a string}
}
\value{
a 1x tibble of the most recent recommendation
}
\description{
Get the most recent recommendation
}
