% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving-average.R
\name{moving_avg}
\alias{moving_avg}
\title{Take moving average of time series data}
\usage{
moving_avg(dat, k)
}
\arguments{
\item{dat}{the data set to transform}

\item{k}{size of the moving average window}
}
\value{
a well formated tibble
}
\description{
Take moving average of time series data
}
