% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-eval.R
\name{new_eval}
\alias{new_eval}
\title{Create a new eval from template}
\usage{
new_eval(ticker, date = Sys.Date())
}
\arguments{
\item{ticker}{an equity security}

\item{date}{the eval date}
}
\value{
a markdown document
}
\description{
Create a new eval from template
}
