% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize data}
\usage{
normalize(dat, norm_date)
}
\arguments{
\item{dat}{a tibble with datatime dates column}

\item{norm_date}{character}
}
\value{
a tibble
}
\description{
Normalize data
}
