% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam-plot.R
\name{pam_plot}
\alias{pam_plot}
\title{Creates a PAM formatted ggplot object}
\usage{
pam_plot(
  p,
  x_lab = NA,
  y_lab = NA,
  plot_title = NA,
  plot_subtitle = NA,
  caption = TRUE,
  axis_titles = FALSE,
  y_format = "comma",
  show_legend = TRUE,
  faceted = NA
)
}
\arguments{
\item{p}{a ggplot object}

\item{x_lab}{string}

\item{y_lab}{string}

\item{plot_title}{string}

\item{plot_subtitle}{string}

\item{caption}{logical}

\item{axis_titles}{logical}

\item{y_format}{string}

\item{show_legend}{logical}

\item{faceted}{string}
}
\value{
a ggplot object
}
\description{
Creates a PAM formatted ggplot object
}
