% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent-change.R
\name{pchange}
\alias{pchange}
\title{Title}
\usage{
pchange(dat, k = 1, trading_days = FALSE)
}
\arguments{
\item{dat}{a tibble}

\item{k}{number of periods over which to calculate the percentage change or a character coercible to a date (YYYY-MM-DD)}

\item{trading_days}{logical. Should compare date be revised to fall on the most recent prior trading day.}
}
\value{
a tibble
}
\description{
Title
}
