% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt-output.R
\name{ppt_output}
\alias{ppt_output}
\title{Title}
\usage{
ppt_output(p, image_name, save = TRUE)
}
\arguments{
\item{p}{a ggplot2 object}

\item{image_name}{character}

\item{save}{logical}
}
\value{
Saves a .png
}
\description{
Title
}
