% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-and-load.R
\name{run_and_load}
\alias{run_and_load}
\title{Title}
\usage{
run_and_load(directory, script, plot_name = NA)
}
\arguments{
\item{directory}{the project name where the script lives}

\item{script}{the script name}

\item{plot_name}{the plot name}
}
\value{
a ggplot2 object
}
\description{
Title
}
