% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-it.R
\name{run_it}
\alias{run_it}
\title{Title}
\usage{
run_it(directory, script, type = "economics")
}
\arguments{
\item{directory}{the project name where the script lives}

\item{script}{the script name}

\item{type}{either "economics" or "equity"}
}
\value{
nothing. just sources the script
}
\description{
Title
}
