% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_layout.R
\name{set_layout}
\alias{set_layout}
\title{Set the layout for a page of the Economic & Market Commentary}
\usage{
set_layout(x)
}
\arguments{
\item{x}{an integer corresponding to the desired layout}
}
\value{
a matrix
}
\description{
Set the layout for a page of the Economic & Market Commentary
}
