% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_title.R
\name{set_title}
\alias{set_title}
\title{Set title of Economic & Market Commentary page}
\usage{
set_title(title)
}
\arguments{
\item{title}{a character}
}
\value{
a text grob
}
\description{
Set title of Economic & Market Commentary page
}
