% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tam-output.R
\name{tam_output}
\alias{tam_output}
\title{Formats charts for use in Tamarac Economic & Market Commentary}
\usage{
tam_output(p, image_name, save = TRUE)
}
\arguments{
\item{p}{ggplot2 object}

\item{image_name}{string}

\item{save}{logical}
}
\value{
ggplot2 object
}
\description{
Formats charts for use in Tamarac Economic & Market Commentary
}
