% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update all data in a directory}
\usage{
update_data(security)
}
\arguments{
\item{security}{Blomberg security ID for equity or economic index}
}
\description{
Update all data in a directory
}
