% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent-change.R
\name{pchange}
\alias{pchange}
\title{Title}
\usage{
pchange(dat, k = 1)
}
\arguments{
\item{dat}{a tibble}

\item{k}{number of periods over which to calculate the percentage change}
}
\value{
a tibble
}
\description{
Title
}
